extract to a separate util
