+++ lightly_studio_view/src/lib/components/SampleDetails/SampleDetailsAnnotationSegment/SampleDetailsAnnotationSegment.helpers.ts
@@ -54,23 +54,6 @@ export const groupAnnotationsBySource = (
 };
 
 /**
- * Counts the distinct annotation sources that have at least one visible (not hidden)
- * annotation.
- */
-export const countVisibleSources = (
-    annotations: AnnotationView[],
-    hiddenAnnotationIds: Set<string>
-): number => {
-    const visibleSourceIds = new Set<string>();
-    for (const annotation of annotations) {
-        if (!hiddenAnnotationIds.has(annotation.sample_id)) {
-            visibleSourceIds.add(annotation.annotation_collection_id);
-        }
-    }
-    return visibleSourceIds.size;
-};
-
-/**
  * Whether every annotation in the list is hidden.
  */
 export const areAllAnnotationsHidden = (
